requirements.txt

flask
tensorflow
opencv-python
pillow
